requirements.txt
mysql-connector-python==9.0.0
pandas==2.2.2
plotly==5.23.0
streamlit==1.37.1
streamlit_navigation_bar==3.3.0
openpyxl==3.1.5
